Some updates from dreamer branch

+++ mbpo/models.py
@@ -36,7 +36,7 @@ class WorldModel(tf.Module):
             loc=self._next_observation_residual_mu(x) + tf.stop_gradient(observation),
             scale_diag=self._next_observation_stddev(x)),
             reward=tfd.Normal(loc=self._reward_mu(x), scale=1.0),
-            terminal=tfd.Bernoulli(logits=self._terminal_logit(x)))
+            terminal=tfd.Bernoulli(logits=self._terminal_logit(x), dtype=tf.float32))
 
 
 class Actor(tf.Module):
